fast login and closable

+++ src/main/java/com/clfang/community/community/dto/QuestionDto.java
@@ -19,7 +19,7 @@ public class QuestionDto {
     private String tag;//标签
     private long gmtCreate;//创建时间
     private long gmtModified;//修改时间
-    private Integer creator;//作者(作者的id)
+    private long creator;//作者(作者的id)
     private  Integer commentCount;//回复数
     private Integer viewCount;//浏览数
     private Integer likeCount;//点赞数
